trying to fix some ts problems
