+++ apps/nextjs/src/pages/_app.tsx
@@ -2,13 +2,14 @@ import "../styles/globals.css";
 
 import type { AppType } from "next/app";
 
+import { AuthContextProvider } from "@struct/auth-context";
+
 import { api } from "~/utils/api";
-import { AuthContextProvider } from "~/authContext";
 
 const MyApp: AppType = ({ Component, pageProps: { ...pageProps } }) => {
   return (
     // <SessionProvider session={session}>
-    <AuthContextProvider>
+    <AuthContextProvider api={api}>
       <Component {...pageProps} />
     </AuthContextProvider>
     // </SessionProvider>
