+++ apps/nextjs/tsconfig.json
@@ -14,6 +14,7 @@
     "**/*.tsx",
     "**/*.cjs",
     "**/*.mjs",
-    ".next/types/**/*.ts"
+    ".next/types/**/*.ts",
+    "../../packages/auth-context/lucia.d.ts"
   ]
 }
