+++ packages/api/src/lucia/lucia.d.ts
@@ -1,6 +1,7 @@
 // lucia.d.ts
 /// <reference types="lucia" />
 declare namespace Lucia {
+  // eslint-disable-next-line @typescript-eslint/consistent-type-imports
   type Auth = import("./index").Auth;
   type UserAttributes = {
     email: string;
