--- packages/auth-context/src/signIn.ts
@@ -1,11 +1,0 @@
-import { Session } from "./context";
-
-export function signInFromContext(
-  setSession: React.Dispatch<React.SetStateAction<Session | null>>,
-) {
-  function signIn() {
-    setSession({});
-  }
-
-  return signIn;
-}
