--- packages/auth-context/src/signOut.ts
@@ -1,11 +1,0 @@
-import { Session } from "./context";
-
-export function signOutFromContext(
-  setSession: React.Dispatch<React.SetStateAction<Session | null>>,
-) {
-  function signOut() {
-    setSession(null);
-  }
-
-  return signOut;
-}
