+++ packages/auth-context/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src", "*.ts", "env.mjs"]
+  "include": ["src", "*.ts", "env.mjs", "./lucia.d.ts"]
 }
